fix: PR 리뷰 피드백 반영 (500 에러, 인증 결함, DIP 위반 해결) - fix: StudyMapper MyBatis 파라미터 바인딩 500 에러 수정 - feat: @LoginUser ArgumentResolver 추가 및 BOLA 취약점 방어 (본인 리소스만 수정/삭제 가능하도록 쿼리 검증 추가) - refactor: PostDAO 인터페이스 분리 및 DIP 원칙 준수하도록 구조 개선 - fix: PostDTO userNum 필드 타입 불일치 오류 수정 (Integer -> Long)

+++ backend/src/main/java/com/provit/dto/community/PostDTO.java
@@ -19,7 +19,7 @@ public class PostDTO {
     // 1. T_POST 테이블 기본 컬럼
     private Long postNum;          // 게시글 고유 번호
     private Integer categoryNum;   // 카테고리 번호
-    private Integer userNum;       // 작성자 회원 번호
+    private Long userNum;       // 작성자 회원 번호
     private String postTitle;      // 게시글 제목
     private String postContent;    // 게시글 본문
     private Integer postLikeCount; // 좋아요 수
